requirements.txt
python-telegram-bot
yfinance
scikit-learn
imbalanced-learn
pandas
numpy